Create movie_data.py
The file contains a sample code which scrapes movie data from "Box Office Mojo" Website. The code is able to scrape data such as total gross earnings for a specified year, movie which was shown in max theaters and total number of theaters. The data is stored in a text file known as movies.txt